feat: refactoring to prepare verify
