+++ pact/ffi/ffi_verifier.py
@@ -1,6 +1,8 @@
 """Wrapper to pact reference dynamic libraries using FFI."""
+from pact.pact_exception import PactException
 from cffi import FFI
 import platform
+
 
 class FFIVerify(object):
     """A Pact Verifier Wrapper."""
@@ -15,9 +17,11 @@ class FFIVerify(object):
         result = lib.pactffi_version()
         return ffi.string(result).decode('utf-8')
 
-
-    def verify(self):
+    def verify(self, *pacts, provider_base_url, provider, enable_pending=False,
+               include_wip_pacts_since=None, **kwargs):
         """Call verify method."""
+        self._validate_input(pacts, **kwargs)
+
         ffi = FFI()
         ffi.cdef("""
         char *pactffi_verify(void);
@@ -25,8 +29,6 @@ class FFIVerify(object):
         lib = self._load_ffi_library(ffi)
         result = lib.pactffi_version()
         return ffi.string(result).decode('utf-8')
-
-# pactffi_verify
 
     def _load_ffi_library(self, ffi):
         """Load the right library."""
@@ -45,3 +47,12 @@ class FFIVerify(object):
             raise Exception(msg)
 
         return ffi.dlopen(libname)
+
+    def _validate_input(self, pacts, **kwargs):
+        if len(pacts) == 0 and not self._broker_present(**kwargs):
+            raise PactException('Pact urls or Pact broker required')
+
+    def _broker_present(self, **kwargs):
+        if kwargs.get('broker_url') is None:
+            return False
+        return True
