+++ requirements_dev.txt
@@ -2,18 +2,19 @@ Click>=2.0.0
 coverage==5.4
 Flask==1.1.1
 configparser==3.5.0
+fastapi==0.67.0
 flake8==3.8.3
 mock==3.0.5
 psutil==5.7.0
 pycodestyle==2.6.0
 pydocstyle==4.0.1
-tox==3.14.0
+tox==3.24.1
 pytest==5.4.1
 pytest-cov==2.11.1
 requests==2.26
-tox-travis==0.8
 urllib3>=1.26.5
 wheel==0.24.0
-requests>=2.5.0
 six>=1.9.0
+fastapi==0.67.0
+uvicorn==0.14.0
 cffi==1.14.6
