+++ tests/test_pact.py
@@ -484,6 +484,7 @@ class PactVerifyTestCase(PactTestCase):
          .upon_receiving('a specific request to the server')
          .with_request('GET', '/path')
          .will_respond_with(200, body='success'))
+
         self.get_verification_call = call(
             'get', 'http://localhost:1234/interactions/verification',
             headers={'X-Pact-Mock-Service': 'true'},
@@ -529,7 +530,6 @@ class PactVerifyTestCase(PactTestCase):
         self.assertEqual(self.mock_requests.call_count, 2)
         self.mock_requests.assert_has_calls([
             self.get_verification_call, self.post_publish_pacts_call])
-
 
 class PactContextManagerTestCase(PactTestCase):
     def setUp(self):
